Update README.md
From Dough to Dollars: Analyzing Maven Pizza's Story

Imagine stepping into a vibrant pizzeria bustling with activity—the aroma of freshly baked dough, the sizzle of melting cheese, and the joyous chatter of satisfied customers. Our mission is to uncover the secrets behind the success of this fictitious pizza haven, Maven Pizza, by delving into a year's worth of sales data. Through this journey, we aim to extract valuable insights to guide business decisions and elevate the pizzeria's operations to new heights.